removing keyboardinputs
